# src/main/resources/swagger.yml
swagger: "2.0"
info:
  description: "Marketplace"
  version: "1.0.0"
  title: "Marketplace"
basePath: "/v1"
...
schemes:
- "https"
- "http"
paths:
...
  /salesman:
    post:
      tags:
      - "salesman"
      summary: "Add a new salesman"
      description: "Add a new salesman"
      operationId: "addSalesman"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "New Salesman"
        required: true
        schema:
          $ref: "#/definitions/SalesmanModel"
      responses:
        "201":
          description: "successful operation"
        "400":
          description: "Invalid product"
  /customer:
    post:
      tags:
      - "customer"
      summary: "Register a new customer"
      description: "Register a new customer"
      operationId: "addCustomer"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "New Customer"
        required: true
        schema:
          $ref: "#/definitions/CustomerModel"
      responses:
        "201":
          description: "successful operation"
        "400":
          description: "Invalid product"
definitions:
...
  CustomerModel:
    type: "object"
    required:
    - name
    properties:
      name:
        type: "string"
  SalesmanModel:
    type: "object"
    required:
    - name
    properties:
      name:
        type: "string"